msm8976-common: OVERRIDE_PRODUCT_COMPRESSED_APEX := false
This was needed for Android 14 QPR1, then removed or changed in QPR2.

However, it doesn't seem to work, so LineageOS-UL hardcoded it at
LineageOS/android_build@1d10d68

Set PRODUCT_COMPRESSED_APEX := false
for some reason we can not disable it from device side.

Just leaving this in case I need it in future.