msm8976-common: add sepolicy for SDK 27 and 29 to allow access to vendor_file
Older apps like F-Droid 1.17 are targeting older SDK 27 and
F-Droid 1.20 (the latest released on May 14, 2024) targeting SDK 29.

With QPR3, it seems that AOSP is restricting and removing some access
in order to "improve security".

So I added sepolicies for apps that target SDK 27 and 29 access to
vendor_file:file with rx_file_perms;

For reference, Android 14 is SDK 34. See

https://apilevels.com/

SDK 27 is Android 8
SDK 29 is Android 10

NOTE: I'm NOT writing a sepolicy for just the F-Droid app. I'm writing
a sepolicy for ALL apps that target SDK 27 and 29. There's no way I
could write a sepolicy for thousands of apps I never use.
So this sepolicy is "generic".